image-generation/requirements.txt
gradio
pillow
torch
transformers
peft
protobuf
python-dotenv
torchvision
sentencepiece